pyproject.toml
[tool.poetry]
name = "anel"
version = "0.1.0"
description = ""
authors = ["ernest <shklyar.ed@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-jose = {version="^3.3.0", extras=["cryptography"]}
passlib = {version="^1.7.4", extras=["bcrypt"]}
pydantic = "^2.0.2"
pydantic-settings = "^2.0.1"

[tool.poetry.dev-dependencies]
pylint = "^2.17.4"

[tool.black]
line-length = 120

[tool.pylint.basic]
extension-pkg-whitelist = "pydantic"

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "import-error"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
